+++ engine/init.lua
@@ -1,0 +1,5 @@
+local path = ... .. "."
+
+require (path .. "reader")
+
+return require (path .. "engine")
